llm-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
pytesseract==0.3.10
Pillow==10.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
requests-mock==1.11.0
python-dotenv==1.0.0
